fact-checker/backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
python-dotenv
langchain-google-genai
